third_party/blink/web_tests/platform/generic/fast/dom/Window/Location/location-override-valueOf-expected.txt

PASS result is correctValue
PASS result is correctValue
PASS (function() { 'use strict'; location.valueOf = 'haxored'; return location.valueOf(); }()) threw exception TypeError: Cannot assign to read only property 'valueOf' of object '[object Location]'.
PASS (function() { location.valueOf = 'haxored'; return location.valueOf(); }()) did not throw exception.
PASS (function() { 'use strict'; delete location.valueOf; }()) threw exception TypeError: Cannot delete property 'valueOf' of [object Location].
PASS (function() { delete location.valueOf; return !!location.valueOf}()) is true
PASS successfullyParsed is true

TEST COMPLETE
